Refactor: Relocate BLoC-related files and update imports
This commit reorganizes the project structure by moving BLoC (Business Logic Component) related files to a dedicated `blocs` directory at the root of the `lib` folder. This improves the separation of concerns and makes the project structure more aligned with common Flutter/Dart practices.

Specifically, the following files and directories have been moved:
- `lib/presentation/app/bloc/` moved to `lib/blocs/app/`
- `lib/presentation/trip/bloc/` moved to `lib/blocs/trip/`

This includes the following key files:
- `MasterPageBloc`, `MasterPageEvents`, `MasterPageStates`
- `TripBloc`, `TripEvents`, `TripStates`
- `BlocProviderExt` (renamed from `BlocExtensions.dart` to `bloc_extensions.dart` for consistency)

All import paths across the codebase have been updated to reflect these changes. This primarily affects files within the `lib/presentation/` directory that were previously importing BLoC components from their old locations.
