+++ lib/blocs/app/master_page_events.dart
