+++ lib/blocs/bloc_extensions.dart
@@ -1,10 +1,10 @@
 import 'package:flutter/material.dart';
 import 'package:flutter_bloc/flutter_bloc.dart';
-import 'package:wandrr/presentation/trip/bloc/bloc.dart';
-import 'package:wandrr/presentation/trip/bloc/events.dart';
+import 'package:wandrr/blocs/trip/bloc.dart';
+import 'package:wandrr/blocs/trip/events.dart';
 
-import 'master_page_bloc.dart';
-import 'master_page_events.dart';
+import 'app/master_page_bloc.dart';
+import 'app/master_page_events.dart';
 
 extension BlocProviderExt on BuildContext {
   void addAuthenticationEvent(AuthenticationEvent event) {
