+++ lib/blocs/trip/bloc.dart
