+++ lib/blocs/trip/events.dart
