+++ lib/blocs/trip/states.dart
